Revert "Merge branch 'GamesLab'"
This reverts commit 3761d97e98cb86c2048b0bd8046393a069d9ed2b, reversing
changes made to 80534522eeee876ee582744b0036ef8f494eec83.
