+++ AsteroidsEntityFramework/Asteroid/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IEntityProcessingService
@@ -1,2 +1,2 @@
-dk.sdu.mmmi.cbse.asteroidSystem.AsteroidControlSystem
-dk.sdu.mmmi.cbse.asteroidSystem.AsteroidSplitter
+dk.sdu.mmmi.cbse.asteroid.AsteroidControlSystem
+dk.sdu.mmmi.cbse.asteroid.AsteroidSplitter
