--- AsteroidsEntityFramework/Bullet/src/main/java/dk/sdu/mmmi/cbse/bulletSystem/Bullet.java
@@ -1,9 +1,0 @@
-package dk.sdu.mmmi.cbse.bulletSystem;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-
-public class Bullet extends Entity {
-
-
-
-}
