--- AsteroidsEntityFramework/Bullet/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IGamePluginService
@@ -1,1 +1,0 @@
-dk.sdu.mmmi.cbse.bulletSystem.BulletPlugin
