+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/data/entityparts/MovingPart.java
@@ -76,7 +76,7 @@ public class MovingPart
             radians -= rotationSpeed * dt;
         }
 
-        // accelerating
+        // accelerating            
         if (up) {
             dx += cos(radians) * acceleration * dt;
             dy += sin(radians) * acceleration * dt;
