+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/data/entityparts/SplitterPart.java
@@ -3,7 +3,7 @@ package dk.sdu.mmmi.cbse.common.data.entityparts;
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 
-public class SplitterPart implements EntityPart {
+public class SplitterPart implements EntityPart{
 
     boolean splitAsteroid = false;
