--- AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/services/IBulletPluginService.java
@@ -1,14 +1,0 @@
-package dk.sdu.mmmi.cbse.common.services;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-import dk.sdu.mmmi.cbse.common.data.GameData;
-
-public interface IBulletPluginService {
-
-     Entity start(Entity entity, GameData gameData);
-
-
-     Entity stop();
-
-
-}
