--- AsteroidsEntityFramework/CommonAsteroid/src/main/java/dk/sdu/mmmi/cbse/commonAsteroid/Asteroid.java
@@ -1,7 +1,0 @@
-package dk.sdu.mmmi.cbse.commonAsteroid;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-
-public class Asteroid extends Entity {
-
-}
