--- AsteroidsEntityFramework/CommonAsteroid/src/main/java/dk/sdu/mmmi/cbse/commonAsteroid/IAsteroidSplitter.java
@@ -1,10 +1,0 @@
-package dk.sdu.mmmi.cbse.commonAsteroid;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-import dk.sdu.mmmi.cbse.common.data.World;
-
-public interface IAsteroidSplitter {
-
-    void createSplitAsteroid(Entity entity, World world);
-
-}
