+++ AsteroidsEntityFramework/Core/src/main/java/dk/sdu/mmmi/cbse/main/Game.java
@@ -5,31 +5,29 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
-import dk.sdu.mmmi.cbse.bulletSystem.BulletControlSystem;
-import dk.sdu.mmmi.cbse.collisionSystem.CollisionSystem;
-import dk.sdu.mmmi.cbse.asteroidSystem.AsteroidControlSystem;
-import dk.sdu.mmmi.cbse.asteroidSystem.AsteroidPlugin;
+import dk.sdu.mmmi.cbse.asteroid.AsteroidControlSystem;
+import dk.sdu.mmmi.cbse.asteroid.AsteroidPlugin;
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
 import dk.sdu.mmmi.cbse.common.services.IEntityProcessingService;
 import dk.sdu.mmmi.cbse.common.services.IGamePluginService;
-import dk.sdu.mmmi.cbse.common.services.IPostEntityProcessingService;
-import dk.sdu.mmmi.cbse.enemySystem.EnemyControlSystem;
-import dk.sdu.mmmi.cbse.enemySystem.EnemyPlugin;
+import dk.sdu.mmmi.cbse.enemysystem.EnemyControlSystem;
+import dk.sdu.mmmi.cbse.enemysystem.EnemyPlugin;
 import dk.sdu.mmmi.cbse.managers.GameInputProcessor;
-import dk.sdu.mmmi.cbse.playerSystem.PlayerControlSystem;
-import dk.sdu.mmmi.cbse.playerSystem.PlayerPlugin;
+import dk.sdu.mmmi.cbse.playersystem.PlayerControlSystem;
+import dk.sdu.mmmi.cbse.playersystem.PlayerPlugin;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Game implements ApplicationListener {
+public class Game
+        implements ApplicationListener {
 
     private static OrthographicCamera cam;
     private ShapeRenderer sr;
+
     private final GameData gameData = new GameData();
     private List<IEntityProcessingService> entityProcessors = new ArrayList<>();
-    private List<IPostEntityProcessingService> entityPostProcessorServiceList = new ArrayList<>();
     private List<IGamePluginService> entityPlugins = new ArrayList<>();
     private World world = new World();
 
@@ -49,45 +47,24 @@ public class Game implements ApplicationListener {
                 new GameInputProcessor(gameData)
         );
 
-
         IGamePluginService playerPlugin = new PlayerPlugin();
-        IEntityProcessingService playerProcess = new PlayerControlSystem();
-
         IGamePluginService enemyPlugin = new EnemyPlugin();
+        IGamePluginService asteroidPlugin = new AsteroidPlugin();
+        IEntityProcessingService playerProcess = new PlayerControlSystem();
         IEntityProcessingService enemyProcess = new EnemyControlSystem();
-
-        IGamePluginService asteroidPlugin = new AsteroidPlugin(3);
         IEntityProcessingService asteroidProcess = new AsteroidControlSystem();
 
-        IPostEntityProcessingService collisionProcess = new CollisionSystem();
-
-        // bullet instantiation here
-
-        IEntityProcessingService bulletProcess = new BulletControlSystem();
-
-
-        // Adding the entities to the game
         entityPlugins.add(playerPlugin);
-       // entityPlugins.add(bulletPlugin);
-        entityPlugins.add(enemyPlugin);
-        entityPlugins.add(asteroidPlugin);
-
-        // Adding the controlsystems to the game
         entityProcessors.add(playerProcess);
+        entityPlugins.add(enemyPlugin);
         entityProcessors.add(enemyProcess);
+        entityPlugins.add(asteroidPlugin);
         entityProcessors.add(asteroidProcess);
-        entityProcessors.add(bulletProcess);
-
-        // Adding the processor of collision to the game
-        entityPostProcessorServiceList.add(collisionProcess);
-
-        // bullet instance here
 
         // Lookup all Game Plugins using ServiceLoader
         for (IGamePluginService iGamePlugin : entityPlugins) {
             iGamePlugin.start(gameData, world);
         }
-
     }
 
     @Override
@@ -111,35 +88,22 @@ public class Game implements ApplicationListener {
         for (IEntityProcessingService entityProcessorService : entityProcessors) {
             entityProcessorService.process(gameData, world);
         }
-
-        for (IPostEntityProcessingService entityPostProcessorService : entityPostProcessorServiceList) {
-            entityPostProcessorService.process(gameData, world);
-        }
-
-
-
-
-
     }
 
     private void draw() {
-        // the draw-method uses the shape-renderer
-        // calculates the shape
         for (Entity entity : world.getEntities()) {
 
             sr.setColor(1, 1, 1, 1);
 
             sr.begin(ShapeRenderer.ShapeType.Line);
-            // sr draws the line
-            // the update method calculates the shape based on an array of vectors
 
             float[] shapex = entity.getShapeX();
             float[] shapey = entity.getShapeY();
-            // shapex = array of vectors
+
             for (int i = 0, j = shapex.length - 1;
                     i < shapex.length;
                     j = i++) {
-                // draw line between the vectors so the vector gets drawn on the screen
+
                 sr.line(shapex[i], shapey[i], shapex[j], shapey[j]);
             }
