+++ AsteroidsEntityFramework/Enemy/src/main/java/dk/sdu/mmmi/cbse/enemysystem/Enemy.java
@@ -1,4 +1,4 @@
-package dk.sdu.mmmi.cbse.enemySystem;
+package dk.sdu.mmmi.cbse.enemysystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
