+++ AsteroidsEntityFramework/Enemy/src/main/java/dk/sdu/mmmi/cbse/enemysystem/EnemyControlSystem.java
@@ -1,9 +1,8 @@
-package dk.sdu.mmmi.cbse.enemySystem;
+package dk.sdu.mmmi.cbse.enemysystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
-import dk.sdu.mmmi.cbse.common.data.entityparts.LifePart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.MovingPart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.PositionPart;
 import dk.sdu.mmmi.cbse.common.services.IEntityProcessingService;
@@ -11,12 +10,9 @@ import dk.sdu.mmmi.cbse.common.services.IEntityProcessingService;
 public class EnemyControlSystem implements IEntityProcessingService {
     @Override
     public void process(GameData gameData, World world) {
-
         for (Entity enemy : world.getEntities(Enemy.class)) {
-
             PositionPart positionPart = enemy.getPart(PositionPart.class);
             MovingPart movingPart = enemy.getPart(MovingPart.class);
-            LifePart lifePart = enemy.getPart(LifePart.class);
 
             movingPart.setLeft(false);
             movingPart.setRight(true);
@@ -25,7 +21,6 @@ public class EnemyControlSystem implements IEntityProcessingService {
 
             movingPart.process(gameData, enemy);
             positionPart.process(gameData, enemy);
-            lifePart.process(gameData,enemy);
 
             updateShape(enemy);
         }
@@ -54,7 +49,6 @@ public class EnemyControlSystem implements IEntityProcessingService {
 
         enemy.setShapeX(shapex);
         enemy.setShapeY(shapey);
-
     }
