+++ AsteroidsEntityFramework/Enemy/src/main/java/dk/sdu/mmmi/cbse/enemysystem/EnemyPlugin.java
@@ -1,30 +1,29 @@
-package dk.sdu.mmmi.cbse.playerSystem;
+package dk.sdu.mmmi.cbse.enemysystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
-import dk.sdu.mmmi.cbse.common.data.entityparts.LifePart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.MovingPart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.PositionPart;
 import dk.sdu.mmmi.cbse.common.services.IGamePluginService;
 
-public class PlayerPlugin implements IGamePluginService {
+public class EnemyPlugin implements IGamePluginService {
 
-    private Entity player;
+    private Entity enemy;
 
-    public PlayerPlugin() {
+    public EnemyPlugin () {
+
     }
 
     @Override
     public void start(GameData gameData, World world) {
-        
-        // Add entities to the world
-        player = createPlayerShip(gameData);
-        world.addEntity(player);
+        // Add enemies to the world
+        enemy = createEnemyShip(gameData);
+        world.addEntity(enemy);
 
     }
 
-    private Entity createPlayerShip(GameData gameData) {
+    private Entity createEnemyShip(GameData gameData) {
 
         float deacceleration = 10;
         float acceleration = 200;
@@ -34,24 +33,16 @@ public class PlayerPlugin implements IGamePluginService {
         float y = gameData.getDisplayHeight() / 2;
         float radians = 3.1415f / 2;
 
-        
-        Entity playerShip = new Player();
+        Entity enemyShip = new Enemy();
+        enemyShip.add(new MovingPart(deacceleration, acceleration, maxSpeed, rotationSpeed));
+        enemyShip.add(new PositionPart(x, y, radians));
 
-
-
-        playerShip.add(new MovingPart(deacceleration, acceleration, maxSpeed, rotationSpeed));
-        playerShip.add(new PositionPart(x, y, radians));
-        playerShip.add(new LifePart(1,1000));
-
-
-        
-        return playerShip;
+        return enemyShip;
     }
 
     @Override
     public void stop(GameData gameData, World world) {
-        // Remove entities
-        world.removeEntity(player);
+        // Remove enemies
+        world.removeEntity(enemy);
     }
-
 }
