+++ AsteroidsEntityFramework/Enemy/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IEntityProcessingService
@@ -1,1 +1,1 @@
-dk.sdu.mmmi.cbse.enemySystem.EnemyControlSystem
+dk.sdu.mmmi.cbse.enemysystem.EnemyControlSystem
