+++ AsteroidsEntityFramework/Enemy/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IGamePluginService
@@ -1,1 +1,1 @@
-dk.sdu.mmmi.cbse.enemySystem.EnemyPlugin
+dk.sdu.mmmi.cbse.enemysystem.EnemyPlugin
