+++ AsteroidsEntityFramework/Player/src/main/java/dk/sdu/mmmi/cbse/playersystem/PlayerControlSystem.java
@@ -1,50 +1,40 @@
-package dk.sdu.mmmi.cbse.playerSystem;
+package dk.sdu.mmmi.cbse.playersystem;
 
-import dk.sdu.mmmi.cbse.bulletSystem.BulletPlugin;
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
-
+import static dk.sdu.mmmi.cbse.common.data.GameKeys.LEFT;
+import static dk.sdu.mmmi.cbse.common.data.GameKeys.RIGHT;
+import static dk.sdu.mmmi.cbse.common.data.GameKeys.UP;
 import dk.sdu.mmmi.cbse.common.data.World;
-import dk.sdu.mmmi.cbse.common.data.entityparts.LifePart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.MovingPart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.PositionPart;
 import dk.sdu.mmmi.cbse.common.services.IEntityProcessingService;
-
-import static dk.sdu.mmmi.cbse.common.data.GameKeys.*;
 
 /**
  *
  * @author jcs
  */
 public class PlayerControlSystem implements IEntityProcessingService {
+
     @Override
     public void process(GameData gameData, World world) {
 
         for (Entity player : world.getEntities(Player.class)) {
             PositionPart positionPart = player.getPart(PositionPart.class);
             MovingPart movingPart = player.getPart(MovingPart.class);
-            LifePart lifePart = player.getPart(LifePart.class);
 
             movingPart.setLeft(gameData.getKeys().isDown(LEFT));
             movingPart.setRight(gameData.getKeys().isDown(RIGHT));
             movingPart.setUp(gameData.getKeys().isDown(UP));
-
-            if (gameData.getKeys().isDown(SPACE)) {
-                BulletPlugin bulletPlugin = new BulletPlugin();
-                world.addEntity(bulletPlugin.start(player,gameData));
-
-            }
-
+            
+            
             movingPart.process(gameData, player);
             positionPart.process(gameData, player);
-            lifePart.process(gameData, player);
 
-
-          //  bulletPart.start(player,gameData);
             updateShape(player);
-
         }
     }
+
     private void updateShape(Entity entity) {
         float[] shapex = entity.getShapeX();
         float[] shapey = entity.getShapeY();
@@ -68,4 +58,5 @@ public class PlayerControlSystem implements IEntityProcessingService {
         entity.setShapeX(shapex);
         entity.setShapeY(shapey);
     }
+
 }
