+++ AsteroidsEntityFramework/Player/src/main/java/dk/sdu/mmmi/cbse/playersystem/PlayerPlugin.java
@@ -1,30 +1,28 @@
-package dk.sdu.mmmi.cbse.enemySystem;
+package dk.sdu.mmmi.cbse.playersystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
-import dk.sdu.mmmi.cbse.common.data.entityparts.LifePart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.MovingPart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.PositionPart;
 import dk.sdu.mmmi.cbse.common.services.IGamePluginService;
 
-public class EnemyPlugin implements IGamePluginService {
+public class PlayerPlugin implements IGamePluginService {
 
-    private Entity enemy;
+    private Entity player;
 
-    public EnemyPlugin () {
-
+    public PlayerPlugin() {
     }
 
     @Override
     public void start(GameData gameData, World world) {
-        // Add enemies to the world
-        enemy = createEnemyShip(gameData);
-        world.addEntity(enemy);
-
+        
+        // Add entities to the world
+        player = createPlayerShip(gameData);
+        world.addEntity(player);
     }
 
-    private Entity createEnemyShip(GameData gameData) {
+    private Entity createPlayerShip(GameData gameData) {
 
         float deacceleration = 10;
         float acceleration = 200;
@@ -33,18 +31,18 @@ public class EnemyPlugin implements IGamePluginService {
         float x = gameData.getDisplayWidth() / 2;
         float y = gameData.getDisplayHeight() / 2;
         float radians = 3.1415f / 2;
-
-        Entity enemyShip = new Enemy();
-        enemyShip.add(new MovingPart(deacceleration, acceleration, maxSpeed, rotationSpeed));
-        enemyShip.add(new PositionPart(x, y, radians));
-        enemyShip.add(new LifePart(1,1000));
-
-        return enemyShip;
+        
+        Entity playerShip = new Player();
+        playerShip.add(new MovingPart(deacceleration, acceleration, maxSpeed, rotationSpeed));
+        playerShip.add(new PositionPart(x, y, radians));
+        
+        return playerShip;
     }
 
     @Override
     public void stop(GameData gameData, World world) {
-        // Remove enemies
-        world.removeEntity(enemy);
+        // Remove entities
+        world.removeEntity(player);
     }
+
 }
