+++ AsteroidsEntityFramework/Player/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IGamePluginService
@@ -1,1 +1,1 @@
-dk.sdu.mmmi.cbse.playerSystem.PlayerPlugin
+dk.sdu.mmmi.cbse.playersystem.PlayerPlugin
