+++ AsteroidsEntityFramework/pom.xml
@@ -12,8 +12,6 @@
         <module>Enemy</module>
         <module>Asteroid</module>
         <module>Collision</module>
-        <module>Bullet</module>
-        <module>CommonAsteroid</module>
     </modules>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
